refactor(projects): изменить структуру для формы
При добавлении новых форм и компонентов не будет захламлять рутовую структуру

Jira issue: VEGA-408
