+++ src/ui/features/projects/ProjectForm/steps/DescriptionStep/index.ts
@@ -1,0 +1,1 @@
+export { DescriptionStep } from './DescriptionStep';
