+++ src/ui/features/projects/ProjectForm/steps/DocumentStep/SidebarForm/index.ts
@@ -1,0 +1,2 @@
+export { SidebarForm } from './SidebarForm';
+export { useFileList } from './use-file-list';
