+++ src/ui/features/projects/ProjectForm/steps/DocumentStep/SidebarForm/utils.ts
@@ -1,4 +1,4 @@
-import { getPluralForm } from '../../../../utils/getPluralForm';
+import { getPluralForm } from '../../../../../../../utils/getPluralForm';
 
 type UploadStatisticsType = {
   filesLeftMessage: string;
