+++ src/ui/features/projects/ProjectForm/steps/DocumentStep/index.ts
@@ -1,0 +1,1 @@
+export { DocumentStep } from './DocumentStep';
