+++ src/ui/features/projects/ProjectForm/steps/ParticipantStep/index.ts
@@ -1,0 +1,1 @@
+export { ParticipantStep } from './ParticipantStep';
